rename migration files
Initially the first migration file being run was the *-create-team file.
The problem with this is that 'Teams' table references 'Users',
which hasn't been created at this pointing, leading to a sequelize error.
To solve this I rename the *-create-user file to make it the first
migration file to be run.